+++ CMakeLists_files.cmake
@@ -339,13 +339,11 @@ list (APPEND PUBLIC_HEADER_FILES
 	opm/core/utility/SparseTable.hpp
 	opm/core/utility/SparseVector.hpp
 	opm/core/utility/Spline.hpp
-	opm/core/utility/SplineCommon_.hpp
 	opm/core/utility/StopWatch.hpp
 	opm/core/utility/TridiagonalMatrix.hpp
 	opm/core/utility/UniformTableLinear.hpp
 	opm/core/utility/Units.hpp
 	opm/core/utility/Unused.hpp
-	opm/core/utility/VariableLengthSpline_.hpp
 	opm/core/utility/VelocityInterpolation.hpp
 	opm/core/utility/WachspressCoord.hpp
 	opm/core/utility/buildUniformMonotoneTable.hpp
